Breaking Search and IncSearch highlighting apart.
For regular searching, it works to have both Search and IncSearch set to
the same color because the cursor follows the current term. The issues
comes when running a command like:

    %s/string1/string2/gc

The cursor is at the input prompt at the bottom (prompting you to accept
the substitution or not), so if there are multiple search matches
on-screen, you don't know which one is the 'current' match that you are
being prompted to accept or reject.